Move ConvertTool to another file
