+++ src/PriceListTools/PriceListTool.cs
@@ -2,31 +2,9 @@
 using Microsoft.Office.Interop.Excel;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace RehauSku.PriceListTools
 {
-    internal class ConvertTool : PriceListTool
-    {
-        private Source Current;
-
-        public void GetCurrent()
-        {
-            Current = new Source(ExcelApp.ActiveWorkbook);
-        }
-
-        public void FillTarget()
-        {
-            ExcelApp.ScreenUpdating = false;            
-            FillColumn(Current.SkuAmount, TargetFile.amountCell.Column);
-            FilterByAmount();
-            ExcelApp.ScreenUpdating = true;
-
-            Forms.Dialog.SaveWorkbookAs();
-        }
-    }
-
-
     internal abstract class PriceListTool
     {
         protected private Application ExcelApp = (Application)ExcelDnaUtil.Application;
